+++ methods/createSession.ts
