+++ methods/deleteSession.ts
