+++ methods/fetchSession.ts
