+++ methods/getAuthHook.ts
