+++ methods/getEndpoints.ts
