+++ methods/updateSession.ts
