+++ package.json
@@ -8,10 +8,9 @@
   "license": "MIT",
   "author": "macropygia",
   "files": [
-    "**/*.ts",
+    "*/**/*.ts",
     "tsconfig.json",
     "!**/*.test.ts",
-    "!**/*.config.ts",
     "!**/.*",
     "!**/__*__"
   ],
